refactor: 메서드명, 패키지 수정
