+++ backend/kirikiri/src/main/java/co/kirikiri/persistence/dto/RoadmapFilterType.java
@@ -1,4 +1,4 @@
-package co.kirikiri.domain.roadmap.dto;
+package co.kirikiri.persistence.dto;
 
 public enum RoadmapFilterType {
     LATEST, GOAL_ROOM_COUNT, PARTICIPANT_COUNT
