+++ backend/kirikiri/src/main/java/co/kirikiri/persistence/roadmap/RoadmapQueryRepository.java
@@ -2,7 +2,7 @@ package co.kirikiri.persistence.roadmap;
 
 import co.kirikiri.domain.roadmap.Roadmap;
 import co.kirikiri.domain.roadmap.RoadmapCategory;
-import co.kirikiri.domain.roadmap.dto.RoadmapFilterType;
+import co.kirikiri.persistence.dto.RoadmapFilterType;
 import java.util.List;
 import java.util.Optional;
 
@@ -10,6 +10,6 @@ public interface RoadmapQueryRepository {
 
     Optional<Roadmap> findRoadmapById(final Long roadmapId);
 
-    List<Roadmap> findRoadmapPagesByCond(final RoadmapCategory category, final RoadmapFilterType orderType,
-                                         final Long lastRoadmapId, final int pageSize);
+    List<Roadmap> findRoadmapsByCond(final RoadmapCategory category, final RoadmapFilterType orderType,
+                                     final Long lastRoadmapId, final int pageSize);
 }
